Rendering improvements, paste on left click, and more
Now, the rendering part is a struct with an implementation to resize, stop and start rendering, and to check if
terminal size is too small for rendering. An implementation for a Web Extension was also made, although it's at
its very start, the core is still being developed. The portion of the screen responsible for rendering the prompt
message was also fixed, before, when resizing the terminal window, if the ratio is small enough, the program
crashes, or leaves artifacts everywhere, in this commit, it's now also fixed. Arguments were implemented,
now you can start without arguments as usual, or enter web monitoring mode with `--monitor`, and listen for
downloads from your web browser to start automatically (yet to be implemented)

There are small changes, as a "Loading" message in the center of the screen when downloading a file,
warnings when a URL is invalid or empty, or response, thread, or I/O errors.

Some changes are still nescessary, and these are the known bugs that might be fixed now:
- Pasting the URL on Windows leaves every character of the content duplicated

And these are the changes to be made:
- Develop web extension
- Web monitoring (for downloads)
- Monitor mode