Implemented extension download support
Implemented downloading files from the internet with monitor mode, the way it works is this: your extension looks for
the localhost websocket to connect on port 6969, and when it connects, it adds a listener for a download event, sending
the URL and the filename. The rust backend (hosting the websocket) listens for packets containing this data in a
predefined format, downloading the file as the said filename, from the said URL.

Cookies will be implemented in the future.

Downloads now will be moved to default Downloads folder set in your operating system, if it can't lookup the default folder,
it will simply not move the file, and will download it in the running folder the app is running on, or in the current folder
from your terminal.